// apps/admin/src/shared/api/generated/ja-korea/schemas/directionsUpdateRequest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export interface DirectionsUpdateRequest {
  /**
     * @minLength 0
     * @maxLength 1000
     */
  koreanAddress?: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  englishAddress?: string;
  /**
     * @minLength 0
     * @maxLength 10000
     */
  kakaoMapHtml?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  phone?: string;
  /**
     * @minLength 0
     * @maxLength 100
     */
  fax?: string;
  /**
     * @minLength 0
     * @maxLength 320
     */
  email?: string;
  version: number;
}